+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/node/AbstractNodeVisitor.java
@@ -158,12 +158,6 @@ public abstract class AbstractNodeVisitor<R, P> implements NodeVisitor<R, P> {
     return visitNode(n, p);
   }
 
-  // Compound assignments
-  @Override
-  public R visitStringConcatenateAssignment(StringConcatenateAssignmentNode n, P p) {
-    return visitNode(n, p);
-  }
-
   // Comparison operations
   @Override
   public R visitLessThan(LessThanNode n, P p) {