+++ framework/src/main/java/org/checkerframework/common/value/ValueTransfer.java
@@ -48,7 +48,6 @@ import org.checkerframework.dataflow.cfg.node.NumericalMultiplicationNode;
 import org.checkerframework.dataflow.cfg.node.NumericalPlusNode;
 import org.checkerframework.dataflow.cfg.node.NumericalSubtractionNode;
 import org.checkerframework.dataflow.cfg.node.SignedRightShiftNode;
-import org.checkerframework.dataflow.cfg.node.StringConcatenateAssignmentNode;
 import org.checkerframework.dataflow.cfg.node.StringConcatenateNode;
 import org.checkerframework.dataflow.cfg.node.StringConversionNode;
 import org.checkerframework.dataflow.cfg.node.StringLiteralNode;
@@ -589,13 +588,6 @@ public class ValueTransfer extends CFTransfer {
       combinedRecAnno = qualHierarchy.greatestLowerBound(oldRecAnno, newRecAnno);
     }
     store.insertValue(receiverJE, combinedRecAnno);
-  }
-
-  @Override
-  public TransferResult<CFValue, CFStore> visitStringConcatenateAssignment(
-      StringConcatenateAssignmentNode n, TransferInput<CFValue, CFStore> p) {
-    TransferResult<CFValue, CFStore> result = super.visitStringConcatenateAssignment(n, p);
-    return stringConcatenation(n.getLeftOperand(), n.getRightOperand(), p, result);
   }
 
   @Override