+++ docs/CHANGELOG.md
@@ -117,8 +117,9 @@ transformations are the same as above, because `@NonNull` is the bottom type,
 
 Deprecated `ElementUtils.getSimpleNameOrDescription()` in favor of `getSimpleDescription()`.
 
-Dataflow Framework: new `ExpressionStatementNode` marks an expression that is
-used as a statement.
+Dataflow Framework:
+ * New `ExpressionStatementNode` marks an expression that is used as a statement.
+ * Removed class `StringConcatenateAssignmentNode`, which is now desugared.
 
 **Closed issues:**